LuneOSWindow: now inherits ApplicationWindow from QuickControls2
The goal here is to automatically inherit the capabilities of a
QtQuickControls2 ApplicationWindow template (dimming, focus,
header/footer...).

The ApplicationWindow type itself, which comes from the LuneOS theme,
will now use the LuneOSWindow type, which itself inherits
T.ApplicationWindow. It automatically sets the type to "Card".

It is now recommended to use ApplicationWindow from QtQuickControls2
(together with LuneOS's theme) instead of LuneOSWindow, though the
latter will still continue working as well as before.

The LuneOS.Application QML test stub is now removed, as the C++ build
should directly be used on desktop.

Signed-off-by: Christophe Chapuis <chris.chapuis@gmail.com>