Qt6: move LuneOS style to the correct folder path
a Qt6 style is an actual QML module, and the folder structure should
match the Qt6 style module name.
Here, the LuneOS style's name is "QtQuick.Controls.LuneOS", and thus
it should be in the QtQuick/Controls folder instead of the hacky
QtQuick/Controls.2 folder that was used before.

This solves the issue with "LuneOS module is not installed" error.

Signed-off-by: Christophe Chapuis <chris.chapuis@gmail.com>